security: Create security_file_vfs_ioctl hook
This LSM hook gets called just before the fs/ioctl.c logic delegates the
requested IOCTL command to the file-specific implementation as implemented by
f_op->unlocked_ioctl (or f_op->ioctl_compat).

It is impractical for LSMs to make security guarantees about these f_op
operations without having intimate knowledge of how they are implemented.

Therefore, depending on the enabled Landlock policy, Landlock aims to block the
calls to filp->f_op->unlocked_ioctl(), but permit the calls to the IOCTL
commands which are already implemented in fs/ioctl.c.

The current call graph is:

  * ioctl syscall
    * security_file_ioctl() LSM hook
    * do_vfs_ioctl() - standard operations
      * file_ioctl() - standard file operations
    * vfs_ioctl()    - delegate to file (called if do_vfs_ioctl() was a no-op)
      * filp->f_op->unlocked_ioctl()

Why not use the existing security_file_ioctl() hook?

With the existing security_file_ioctl() hook, it is technically feasible to
prevent the call to filp->f_op->unlocked_ioctl(), but it would be difficult to
maintain: security_file_ioctl() gets called further up the call stack, so an
implementation of it would need to predict whether the logic further below will decide to call
f_op->unlocked_ioctl().  That can only be done by mirroring the logic in
do_vfs_ioctl() to some extent, and keeping this in sync.

We therefore think that it is cleaner to add a new LSM hook, which gets called
closer to the security boundary which we actually want to block,
filp->f_op->unlocked_ioctl().

Why is it difficult to reason about f_op->unlocked_ioctl()?

The unlocked_ioctl() and ioctl_compat() operations generally do the following
things:

 1. Execute code depending on the IOCTL command number,
    to implement the IOCTL command.

 2. Execute code independent(!) of the IOCTL command number,
    usually to implement common locking and resource allocation behavior.

    Notably, this often happens before(!) the implementation looks
    at the command number.

Due to the number of device drivers in Linux, it is infeasible for LSM
maintainers to keep track of what these implementations do in detail.

Due to 2., even permitting selected IOCTL command numbers to be implemented by
devices would probably be a mistake, because even when a device does not
implement a given IOCTL command, it might still execute code when you try to
call it.

Cc: Paul Moore <paul@paul-moore.com>
Cc: Arnd Bergmann <arnd@arndb.de>
Cc: Christian Brauner <brauner@kernel.org>
Cc: Dave Chinner <david@fromorbit.com>
Cc: Mickaël Salaün <mic@digikod.net>
Signed-off-by: Günther Noack <gnoack@google.com>