landlock: transpose the layer masks data structure
The layer masks data structure tracks the requested but unfulfilled
access rights during an operations security check.  It stores one bit
for each combination of access right and layer index.  If the bit is
set, that access right is not granted (yet) in the given layer and we
have to traverse the path further upwards to grant it.

Previously, the layer masks were stored as arrays mapping from access
right indices to layer_mask_t.  The layer_mask_t value then indicates
all layers in which the given access right is still (tentatively)
denied.

This patch introduces struct layer_access_masks instead: This struct
contains an array with the access_mask_t of each (tentatively) denied
access right in that layer.

The hypothesis of this patch is that this simplifies the code enough
so that the resulting code will run faster:

* We can use bitwise operations in multiple places where we previously
  looped over bits individually with macros.  (Should require less
  branch speculation)

* Code is ~160 lines smaller.

Other noteworthy changes:

* Clarify deny_mask_t and the code assembling it.
  * Document what that value looks like
  * Make writing and reading functions specific to file system rules.
    (It only worked for FS rules before as well, but going all the way
    simplifies the code logic more.)
* In no_more_access(), call a new helper function may_refer(), which
  only solves the asymmetric case.  Previously, the code interleaved
  the checks for the two symmetric cases in RENAME_EXCHANGE.  It feels
  that the code is clearer when renames without RENAME_EXCHANGE are
  more obviously the normal case.

Signed-off-by: Günther Noack <gnoack3000@gmail.com>